Skip creating and cleaning up TPUs if using a prestarted TPU.

+++ perfzero/lib/perfzero/utils.py
@@ -407,6 +407,9 @@ def setup_tpu(parameters):
   Returns:
     True if an error occurs during setup.
   """
+  # Skip creating tpu if using a prestarted tpu.
+  if parameters.get('using_prestarted_tpu') == 'true':
+    return False
   try:
     base_cmd = 'gcloud compute tpus execution-groups create'
     args = [
@@ -437,7 +440,10 @@ def cleanup_tpu(parameters):
   Returns:
     True if an error occurs during cleanup.
   """
-  
+  # Skip cleaning up the tpu if using a prestarted tpu.
+  if parameters.get('using_prestarted_tpu') == 'true':
+    return False
+
   base_cmd = 'gcloud compute tpus execution-groups delete'
 
   args = [
